Notification
Hey there. I am the creator and maintainer of the Mushroom Dream tileset for cataclysm. I really apologize for reaching out to you like this, but there is no other way to contact you via Github.

I was recently made aware that the license that I am releasing the artworks in the tileset are not compatible with some terms of a public repository on Github, and thus so is this fork at fault as well. The tileset and it's contents are still (and will alway be) available on their page, mushroom dream, I just have updated the page to accommodate for the necessary changes and opened a discussion in case there's anything you or anyone else needs to discuss.

Due to the changes made to the original page, updating this fork will not work. The solution i found was to simply delete this fork and make a new one. I really apologize for reaching you out like this, but i found it to be a friendlier way than to issue a DMCA takedown out of the blue for something that was a mistake of mine.

If you have any questions, you can always reach me on the main mushroom dream page.

Apologies once again, 
Albert